refactor(common): create ui and dto sub-packages at common package
