+++ src/main/java/woowacourse/common/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto.response;
+package woowacourse.common.dto;
 
 public class ErrorResponse {
